início da alteração para soft-delete

+++ src/main/java/com/tisv/agef/services/ModeloService.java
@@ -2,6 +2,7 @@ package com.tisv.agef.services;
 
 import com.tisv.agef.domains.Modelo;
 import com.tisv.agef.repositories.ModeloRepository;
+import com.tisv.agef.services.exceptions.ConstraintViolationExceptionOnDelete;
 import com.tisv.agef.services.exceptions.ObjectNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -37,7 +38,18 @@ public class ModeloService {
     }
 
     public void delete(Integer id) {
-        repo.deleteById(id);
+        Modelo modelo = this.find(id);
+
+        if (modelo.getPecaFeira() == null) {
+            repo.deleteById(id);
+
+        } else {
+            throw new ConstraintViolationExceptionOnDelete(
+                    "Não é possivel deletar este modelo enquanto existir uma peça associada a ele" +
+                            "\n" + "Modelo: " + modelo.toString() +
+                            "\n" + "Peça: " + modelo.getPecaFeira().toString()
+            );
+        }
     }
 
     public void update(Integer id, Modelo modeloArg) {
